+++ src/utils/index.ts
@@ -1,5 +1,6 @@
 export * from "./addToRange.js";
 export * from "./dateMatchModifiers.js";
+export * from "./rangeContainsDayOfWeek.js";
 export * from "./rangeIncludesDate.js";
 export * from "./rangeMatchModifiers.js";
 export * from "./typeguards.js";
